Move service to a new home
