+++ source/Messaging.Infrastructure/OutgoingMessages/NewMessageAvailableNotifier.cs
@@ -16,6 +16,7 @@ using System;
 using System.Threading.Tasks;
 using Energinet.DataHub.MessageHub.Client.DataAvailable;
 using Energinet.DataHub.MessageHub.Model.Model;
+using Messaging.Application.Actors;
 using Messaging.Application.Configuration;
 using Messaging.Application.OutgoingMessages;
 using Messaging.Domain.OutgoingMessages;
